requirements.txt
# CJE library (main dependency)
cje-eval>=0.2.5

# Core scientific computing (already in cje-eval, listed for clarity)
numpy>=1.26
pandas>=2.2
scipy>=1.11
scikit-learn>=1.4

# Visualization (optional but needed for plots)
matplotlib>=3.8
seaborn>=0.13.2

# Progress bars
tqdm

# For data generation (optional - only if regenerating dataset)
# Uncomment if you need to regenerate responses:
# transformers>=4.52.2
# fireworks-ai>=0.15.0
